Working on automating syncthing for lumbermill
