+++ cgl/plugins/project_management/syncthing/utils.py
@@ -1,73 +1,7 @@
-import click
 import xml.etree.ElementTree as ET
-import gspread
-import requests
-import os
 import getpass
 import subprocess
-from oauth2client.service_account import ServiceAccountCredentials
-
-
-def authorize_sheets(sheet_name):
-    """
-    Authorizes api calls to the google sheet
-    :param sheet_name: Title of the sheet being accessed
-    :return: A google sheet object
-    """
-    scope = ['https://www.googleapis.com/auth/spreadsheets', 'https://www.googleapis.com/auth/spreadsheets',
-             'https://www.googleapis.com/auth/drive']
-
-    filepath = 'C:\\Users\\Molta\\Desktop\\client.json'
-    creds = ServiceAccountCredentials.from_json_keyfile_name(filepath, scope)
-
-    client = gspread.authorize(creds)
-
-    sheet = client.open(sheet_name).sheet1
-    return sheet
-
-
-def does_id_exist(id, sheet):
-    """
-    Checks to see if ID is already entered into the google sheet
-    :param id: ID to look for in sheet
-    :param sheet: Title of google sheet to search in
-    :return: True if the value is already entered in the sheet, false if the value has no entry yet
-    """
-    sheet = authorize_sheets(sheet)
-    data = sheet.col_values(1)
-    for i in data:
-        if i == id:
-            return True
-    return False
-
-
-def find_empty_row_in_sheet():
-    """
-    Finds the first empty row in the google sheet
-    :return: Integer representation of the first empty row
-    """
-    sheet = authorize_sheets('LONE_COCONUT_SYNC_THING')
-    count = 1
-    data = sheet.col_values(1)
-    for i in data:
-        if i == '':
-            return count
-        count += 1
-    return count
-
-
-def get_sheets_authentication(filepath, client='lone-coconut'):
-    """
-    Gets the json authentication file from amazon s3 and saves it on the local machine at the filepath
-    :param client: The name of the client for the sheet
-    :param filepath: The filepath where the authentication json will be saved
-    :return: Returns the filepath to the local copy of the authentication file
-    """
-    url = 'https://%s.s3.amazonaws.com/sync_thing/lone-coconut-syncthing-0853eb66e60e.json' % client
-    r = requests.get(url, allow_redirects=True)
-    with open(filepath, 'w+') as f:
-        f.write(r.content)
-    return filepath
+import cgl.plugins.google.sheets as G
 
 
 def get_config_path():
@@ -99,36 +33,37 @@ def get_my_device_info():
     filepath = get_config_path()
     tree = ET.parse(filepath)
     root = tree.getroot()
-    my_id = root[0][1].get('id')
+    device_id = root[0][1].get('id')
+    machine_name = ''
 
     for child in root:
-        if child.tag == 'device' and child.get('id') == my_id:
+        if child.tag == 'device' and child.get('id') == device_id:
             machine_name = child.get('name')
             print machine_name
 
-    return {'id': my_id, 'name': machine_name}
+    return {'id': device_id, 'name': machine_name}
 
-def add_device_info_to_sheet(sheet_name):
+
+def add_device_info_to_sheet(sheet):
     """
     Adds current device information to google sheet
+    :param sheet: The sheet object to be edited
     :return:
     """
-    sheet = authorize_sheets(sheet_name)
-    new_row = find_empty_row_in_sheet()
+    new_row = G.find_empty_row_in_sheet(sheet)
     device_dictionary = get_my_device_info()
     sheet.update_cell(new_row, 1, device_dictionary['id'])
     sheet.update_cell(new_row, 2, device_dictionary['name'])
     sheet.update_cell(new_row, 3, getpass.getuser().lower())
 
-def get_all_device_info():
+
+def get_all_device_info(sheet):
     """
     Gets all devices currently entered into syncthing
     :return: Array of dictionaries containing necessary device information
     """
-    sheet = authorize_sheets('LONE_COCONUT_SYNC_THING')
     device_list = []
-    num_of_rows = find_empty_row_in_sheet()
-    print num_of_rows
+    num_of_rows = G.find_empty_row_in_sheet(sheet)
     for row in range(2, num_of_rows):
         if sheet.cell(row,1).value != get_my_device_info()['id']:
             entry = {"id": sheet.cell(row, 1).value, "name": sheet.cell(row,2).value, "user": sheet.cell(row,3).value}
@@ -137,17 +72,15 @@ def get_all_device_info():
     return device_list
 
 
-def add_all_device_id_to_config():
+def add_all_devices_to_config(sheet):
     """
     Add a new device to be synched with in syncthing
     :return:
     """
-    device_list = get_all_device_info()
+    device_list = get_all_device_info(sheet)
     filepath = get_config_path()
     tree = ET.parse(filepath)
     root = tree.getroot()
-
-    print root[1]
 
     for entry in device_list:
         new_node = ET.SubElement(root, 'device')
@@ -188,13 +121,18 @@ def add_folder_to_config(folder_id, filepath):
     new_node = ET.SubElement(root, 'folder')
     new_node.set('id', folder_id)
     new_node.set('path', filepath)
+    new_node.set('type', 'sendreceive')
+    new_node.set('rescanIntervalS', '3600')
+    new_node.set('fsWatcherEnabled', "True")
+    new_node.set('fsWatcherDelayS', "10")
+    new_node.set('ignorePerms', "false")
+    new_node.set('autoNormalize', "true")
     tree.write(config_path)
 
 
-
-
 if __name__ =="__main__":
-    # get_sheets_authentication(filepath)
-    # add_device_info_to_sheet(filepath)
-    # add_all_device_id_to_config()
-    # add_folder_to_config('kyls_new_file', 'C:\\Users\\Molta\\test')
+    # file_location = G.get_sheets_authentication('C:\\Users\\Molta\\Desktop')
+    sheet1 = G.authorize_sheets('LONE_COCONUT_SYNC_THING', 'C:\\Users\\Molta\\Desktop\\client.json')
+    # add_device_info_to_sheet(sheet1)
+    #add_all_devices_to_config(sheet1)
+    add_folder_to_config('kyls_new_file', 'C:\\Users\\Molta\\test')
